Corrected some solutions of Chapter 1-8

+++ Chapter 1-8/Solutions/3-calculator.py
@@ -2,10 +2,11 @@
 Basic solution for the calculator
 """
 
+# robust solutions for reading floats will be discussed later in the course
 # read a first number
-x = int(input("Enter a first number:  "))
+x = float(input("Enter a first number:  "))
 # read a second number
-y = int(input("Enter a second number:  "))
+y = float(input("Enter a second number:  "))
 # read an operation
 operation = input("Enter your operation: *  /  +  -  ^  %  //       ")
